scrapper/requirements.txt
selenium==3.141.0
urllib3==1.25.9
redis==3.5.3
python-dotenv==0.13.0
